+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog for NeoFS Node
 ### Changed
 - `object search` CLI command is now the same as `object searchv2` (#3931)
 - Optimized integer comparison in SearchV2 (#3938)
+- Optimized number of nodes for SEARCH in EC containers (#3940)
 
 ### Removed
 - `policer.max_workers` configuration (#3920)